# Small optimization of index_add with dim>1 on the cpu with ~5% speedup